requirements.txt
SpeechRecognition==3.10.0
edge-tts==7.2.8
pygame==2.6.1
pyaudio==0.2.13
requests==2.31.0
beautifulsoup4==4.12.3
psutil==5.9.8
pycaw==20240210
comtypes==1.4.1
matplotlib==3.8.3
numpy==1.26.4
python-docx==1.1.0
openpyxl==3.1.2
reportlab==4.1.0
lxml==5.1.0
playwright>=1.40.0
pywin32==306
pyautogui>=0.9.54
pytesseract>=0.3.10
mss>=9.0.0
vosk>=0.3.45
keyboard>=0.13.5
pyperclip>=1.8.2
